pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sanskrit-rag"
version = "0.2.0"
description = "Minimal Sanskrit RAG service with CLI and FastAPI surfaces"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Vidhi Saxena" }]
keywords = ["sanskrit", "rag", "llm", "fastapi"]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sentence-transformers>=3.4",
    "faiss-cpu>=1.9",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "llama-cpp-python>=0.3",
    "numpy>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
]

[project.urls]
Homepage = "https://github.com/vidhiisaxena/Sanskrit_RagSystem"
Repository = "https://github.com/vidhiisaxena/Sanskrit_RagSystem"
Issues = "https://github.com/vidhiisaxena/Sanskrit_RagSystem/issues"

[tool.hatch.build.targets.wheel]
packages = ["code"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["code", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
